+++ src/main/java/org/crops/fitserver/domain/auth/dto/request/SocialLoginRequest.java
@@ -1,0 +1,8 @@
+package org.crops.fitserver.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record SocialLoginRequest(
+    @NotNull String code
+) {
+}
