+++ src/main/java/org/crops/fitserver/domain/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package org.crops.fitserver.domain.auth.facade.dto;
+package org.crops.fitserver.domain.auth.dto.response;
 
 import jakarta.validation.constraints.NotNull;
 import org.crops.fitserver.global.jwt.TokenCollection;
