+++ src/main/java/org/crops/fitserver/domain/auth/facade/impl/AuthFacadeImpl.java
@@ -1,14 +1,14 @@
 package org.crops.fitserver.domain.auth.facade.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.crops.fitserver.domain.auth.facade.dto.SocialLoginPageResponse;
+import org.crops.fitserver.domain.auth.dto.response.SocialLoginPageResponse;
 import org.crops.fitserver.domain.auth.service.OAuthService;
 import org.crops.fitserver.domain.auth.service.provider.OAuthServiceProvider;
 import org.crops.fitserver.domain.auth.facade.AuthFacade;
 import org.crops.fitserver.global.jwt.JwtProvider;
 import org.crops.fitserver.global.jwt.TokenInfo;
 import org.crops.fitserver.domain.user.domain.SocialPlatform;
-import org.crops.fitserver.domain.auth.facade.dto.TokenResponse;
+import org.crops.fitserver.domain.auth.dto.response.TokenResponse;
 import org.crops.fitserver.domain.user.domain.User;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
